% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtcc_app.R
\name{dtcc_app}
\alias{dtcc_app}
\title{DTCC-app}
\usage{
dtcc_app(load_source_data = T, pre_loaded_data = NULL)
}
\arguments{
\item{load_source_data}{Spesifiser om data skal lastes fra kilde. Tar T eller F}

\item{pre_loaded_data}{Hvis load_source_data == F, spesifiser navn på et datasett lastet inn med GetDTCCData-funksjonen}
}
\value{

}
\description{
DTCC-app
}
\examples{
\dontrun{
#Start app med innlasting av data
dtcc_app()
# Start app med ferdig datasett
d = GetDTCCData()
dtcc_app(F, pre_loaded_data = d)

}
}
